some trivial comments removed

+++ ocean_data_qc/bokeh_models/bokeh_sources.py
@@ -191,10 +191,10 @@ class BokehSources(Environment):
         '''
         cols = self.env.cruise_data.get_cols_by_attrs(['flag'])
         flag_vals = self.env.cds_df[cols].values.ravel('K')  # ravel('K') to flatten the multidimensional array
-        flag_vals = flag_vals[~np.isnan(flag_vals)]          # remove nan
-        flag_vals = np.unique(flag_vals)                     # select the unique values
-        flag_vals = flag_vals.astype(np.int64)               # convert to integer
-        flag_vals = flag_vals.tolist()                       # convert to python list
+        flag_vals = flag_vals[~np.isnan(flag_vals)]
+        flag_vals = np.unique(flag_vals)
+        flag_vals = flag_vals.astype(np.int64)
+        flag_vals = flag_vals.tolist()
 
         # forcing the basic values
         # TODO: Create a flag form in order to set the flag values by hand
